% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMoves.R
\name{bruteForce}
\alias{bruteForce}
\title{Brute force type algotithm to solve the puzzle game.}
\usage{
bruteForce(...)
}
\value{
Invisible: Solved teh puzzle game - updated all structures.
}
\description{
Brute force type algotithm to solve the puzzle game.
}
\examples{
\dontrun{
# very long calculation
bruteForce()
}
}
\seealso{
\code{\link{makeMove}}.
}
