% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{changePath}
\alias{changePath}
\title{Changes path of solving puzzle.}
\usage{
changePath(newPath)
}
\arguments{
\item{newPath}{An integer, number of the prepared paths
(random path: \code{0}) or vector of fields numbers ordered
in a way which field to start and where to go next...}
}
\value{
Invisible: Changed \code{path}.
}
\description{
Changes path of solving puzzle.
}
