% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{checkBorderKind}
\alias{checkBorderKind}
\title{Check type of the board field.}
\usage{
checkBorderKind(field)
}
\value{
An integer: 1 - up, 2 - right, 3 - down, 4 - left.
}
\description{
Check type of the board field.
}
\examples{
checkBorderKind(2)
}
