% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMoves.R
\name{checkEdges}
\alias{checkEdges}
\title{Checks if edges of new pieces are available.}
\usage{
checkEdges(id, pattern)
}
\arguments{
\item{id}{An integer, a number of a piece.}

\item{pattern}{A vector \code{c(U, R, D, L)} - pattern of a piece.}
}
\value{
Logical.
}
\description{
Checks if edges of new pieces are available.
}
\examples{
checkEdges(9, createPattern(12))
}
