% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{checkFieldType}
\alias{checkFieldType}
\title{Check type of the board field.}
\usage{
checkFieldType(field)
}
\value{
An integer: 1 - middle, 2 - border, 3 - corner.
}
\description{
Check type of the board field.
}
\examples{
checkFieldType(1)
checkFieldType(2)
}
