% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{EternityII}
\alias{EternityII}
\alias{EternityII-package}
\title{EternityII: A brute force type algorithm trying solve this puzzle.}
\description{
Different implementation of a brute force type algorithms solving
EternityII puzzle:
\url{https://en.wikipedia.org/wiki/Eternity_II_puzzle}
}
