% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMoves.R
\name{findMoves}
\alias{findMoves}
\title{Finds possible moves (also checks rotations).}
\usage{
findMoves(field)
}
\arguments{
\item{field}{An integer, a number of a field on the board.}
}
\value{
A vector \code{c(U, R, D, L)} - pattern of a piece.
  Updated structure \code{possibleMoves}.
}
\description{
Finds possible moves (also checks rotations).
}
\examples{
findMoves(120)
}
\seealso{
\code{\link{createPattern}}.
}
