% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initiate or restore the puzzle game - EternityII.}
\usage{
init(restore = TRUE, initPiece = TRUE, tip1 = TRUE, tip2 = TRUE,
  tip3 = TRUE)
}
\arguments{
\item{restore}{Logical (default \code{TRUE}). Start new game
or restore previously saved one.}

\item{initPiece}{Logical (default \code{TRUE}). Add init piece to board.}

\item{tip1}{Logical (default \code{TRUE}). Add tip 1 piece to board.}

\item{tip2}{Logical (default \code{TRUE}). Add tip 2 piece to board.}

\item{tip3}{Logical (default \code{TRUE}). Add tip 3 piece to board.}
}
\value{
Invisible: Initiated puzzle game EternityII.
}
\description{
Initiate or restore the puzzle game - EternityII.
}
\examples{
init()
}
\seealso{
\code{\link{changePath}}.
}
