% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init36}
\alias{init36}
\title{Initiate or restore the clue 1 puzzle game.}
\usage{
init36(restore = TRUE, initPiece = TRUE, tip1 = FALSE, tip2 = FALSE,
  tip3 = FALSE)
}
\arguments{
\item{restore}{Logical (default \code{TRUE}). Start new game
or restore previously saved one.}

\item{initPiece}{Logical (default \code{TRUE}). Add init piece to board.}

\item{tip1}{Logical (default \code{FALSE}). Add tip 1 piece to board.}

\item{tip2}{Logical (default \code{FALSE}). Add tip 2 piece to board.}

\item{tip3}{Logical (default \code{FALSE}). Add tip 3 piece to board.}
}
\value{
Invisible: Initiated puzzle test game.
}
\description{
Initiate or restore the clue 1 puzzle game.
}
\examples{
init36()
}
\seealso{
\code{\link{changePath}}.
}
