% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initRotation}
\alias{initRotation}
\title{Initial rotation of restored \code{pieces}.}
\usage{
initRotation()
}
\value{
Invisible: Rotated pieces.
}
\description{
Initial rotation of restored \code{pieces}.
}
