% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMoves.R
\name{makeMove}
\alias{makeMove}
\title{Makes move forward if possible. Undo previous move if there is no
  possibility to move forward.}
\usage{
makeMove(strategy = "Rnd")
}
\arguments{
\item{strategy}{One of the strategies of chosing \code{id} of the piece
for the next move: Rnd, Min, Max.}
}
\value{
\code{id} of piece added to solution (step forward) or removed
  from the solution (step backward).
}
\description{
Makes move forward if possible. Undo previous move if there is no
  possibility to move forward.
}
\examples{
makeMove()
}
\seealso{
\code{\link{bruteForce}}.
}
