% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{pieces72}
\alias{pieces72}
\title{72 puzzle pieces (clue 2).}
\format{A data.table with 256 rows and 6 variables:
\describe{
  \item{Id}{Piece's number,}
  \item{U}{Up edge,}
  \item{R}{Right edge,}
  \item{D}{Down edge,}
  \item{L}{Left edge,}
  \item{Type}{Piece's type: 1 - middle, 2 - border, 3 - corner,}
  \item{Rotation}{0 - not rotated, 1 - turned by 90 degree,
    2 - turned by 180 degree, 3 - turned by 270 degree.}
}}
\usage{
pieces72
}
\description{
A dataset containing info about pieces' Ids, edges and types.
The value of an edge is decoded: 0 for border and 1-8 for
different symbols.
}
\keyword{datasets}
