% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showBoard}
\alias{showBoard}
\title{Show board.}
\usage{
showBoard()
}
\value{
\code{board} matrix \code{N} by \code{N} with Ids of pieces on fields.
}
\description{
Show board.
}
\examples{
showBoard()
}
\seealso{
\code{\link{showPath}}.
}
