% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showPath}
\alias{showPath}
\title{Show path.}
\usage{
showPath(pth = path)
}
\arguments{
\item{pth}{vector of fields numbers of the board (or subset if initial
or/and tips pieces are used.}
}
\value{
\code{board} matrix \code{N} by \code{N} with numbered fields according to path.
}
\description{
Show path.
}
\examples{
showPath(path)
}
\seealso{
\code{\link{showBoard}}.
}
