% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showPiece}
\alias{showPiece}
\title{Show piece (3x3).}
\usage{
showPiece(id)
}
\arguments{
\item{id}{A number of a piece.}
}
\value{
\code{mat} matrix 3 by 3
}
\description{
Show piece (3x3).
}
\examples{
showPiece(9)
}
\seealso{
\code{\link{showPiecePNG}}.
}
