% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showPiecePNG}
\alias{showPiecePNG}
\title{Show piece (png).}
\usage{
showPiecePNG(id)
}
\arguments{
\item{id}{A number of a piece.}
}
\value{
Invisible. Plot.
}
\description{
Show piece (png).
}
\examples{
showPiecePNG(9)
}
\seealso{
\code{\link{showPiece}}.
}
