% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnBorder}
\alias{turnBorder}
\title{Turn border piece to match proper edge of the border.}
\usage{
turnBorder(id, field)
}
\arguments{
\item{id}{An integer, a number of a piece.}

\item{field}{A border field number.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn border piece to match proper edge of the border.
}
\examples{
showPiece(5)
turnBorder(5, 17)
showPiece(5)
turnBorder(5, 32)
showPiece(5)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
