% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnBorderDown}
\alias{turnBorderDown}
\title{Turn border piece down (D=0).}
\usage{
turnBorderDown(id)
}
\arguments{
\item{id}{An integer, a number of a piece.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn border piece down (D=0).
}
\examples{
showPiece(5)
turnBorderDown(5)
showPiece(5)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
