% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnBorderLeft}
\alias{turnBorderLeft}
\title{Turn border piece left (L=0).}
\usage{
turnBorderLeft(id)
}
\arguments{
\item{id}{An integer, a number of a piece.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn border piece left (L=0).
}
\examples{
showPiece(5)
turnBorderLeft(5)
showPiece(5)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
