% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnBorderRight}
\alias{turnBorderRight}
\title{Turn border piece right (R=0).}
\usage{
turnBorderRight(id)
}
\arguments{
\item{id}{An integer, a number of a piece.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn border piece right (R=0).
}
\examples{
showPiece(5)
turnBorderRight(5)
showPiece(5)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
