% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnBorderUp}
\alias{turnBorderUp}
\title{Turn border piece up (U=0).}
\usage{
turnBorderUp(id)
}
\arguments{
\item{id}{An integer, a number of a piece.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn border piece up (U=0).
}
\examples{
showPiece(5)
turnBorderUp(5)
showPiece(5)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
