% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnCorner}
\alias{turnCorner}
\title{Turn corner piece to match proper corner of the border.}
\usage{
turnCorner(id, field)
}
\arguments{
\item{id}{An integer, a number of a piece.}

\item{field}{A corner field number: \code{1}, \code{sqrt(N)},
\code{N - sqrt(N) + 1}, \code{N}.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn corner piece to match proper corner of the border.
}
\examples{
showPiece(1)
turnCorner(1, 1)
showPiece(1)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
