% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnPiece180}
\alias{turnPiece180}
\title{Turn piece by 180 degree.}
\usage{
turnPiece180(id)
}
\arguments{
\item{id}{An integer, a number of a piece.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn piece by 180 degree.
}
\examples{
showPiece(9)
turnPiece180(9)
showPiece(9)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
