% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnPiece270}
\alias{turnPiece270}
\title{Turn piece by 270 degree.}
\usage{
turnPiece270(id)
}
\arguments{
\item{id}{An integer, a number of a piece.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn piece by 270 degree.
}
\examples{
showPiece(9)
turnPiece270(9)
showPiece(9)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
