% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turnPiece90}
\alias{turnPiece90}
\title{Turn piece by 90 degree.}
\usage{
turnPiece90(id)
}
\arguments{
\item{id}{An integer, a number of a piece.}
}
\value{
Invisible: Turned piece.
}
\description{
Turn piece by 90 degree.
}
\examples{
showPiece(9)
turnPiece90(9)
showPiece(9)
}
\seealso{
\code{\link{showPiece}}, \code{\link{showPiecePNG}}.
}
