% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMoves.R
\name{updateEdges}
\alias{updateEdges}
\title{Updates available edges counters.}
\usage{
updateEdges(id, pattern = rep(-1L, 4), add = TRUE)
}
\arguments{
\item{id}{An integer, a number of a piece.}

\item{pattern}{A vector \code{c(U, R, D, L)} - pattern of a piece.}

\item{add}{Logical (default \code{TRUE}), update edges after adding
or removing piece from the board.}
}
\value{
Invisible: Updated \code{usedEdges}
}
\description{
Updates available edges counters.
}
\examples{
\dontrun{
updateEdges(9)
}
}
